Mini-Projects/Breast-Cancer-Best-Model/requirements.txt
pycaret
django-environ>=0.8.1
django-import_export
virtualenv
django-gunicorn
